api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.126.0",
    "imbalanced-learn==0.11.0",
    "joblib>=1.5.3",
    "numpy==1.26.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-decouple>=3.8",
    "python-multipart>=0.0.21",
    "ruff>=0.14.10",
    "scikit-learn==1.3.2",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.2",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["app*","scripts*", "training*", "data*", "testing*"]

[project.scripts]
server = "scripts.server:start"
train = "scripts.train:train"
test = "scripts.test:test"
